Changed health check to indicate HTTPS endpoint connectivity using TLS 1.2
