+++ src/Umbraco.Web.UI/umbraco/config/lang/fr.xml
@@ -2170,9 +2170,6 @@ Pour gérer votre site, ouvrez simplement le backoffice Umbraco et commencez à 
     <key alias="notificationEmailsCheckErrorMessage"><![CDATA[L'adresse email de notification est toujours à sa valeur par défaut : <strong>%0%</strong>.]]></key>
     <key alias="scheduledHealthCheckEmailBody"><![CDATA[<html><body><p>Les résultats de l'exécution du Umbraco Health Checks planifiée le %0% à %1% sont les suivants :</p>%2%</body></html>]]></key>
     <key alias="scheduledHealthCheckEmailSubject">Statut du Umbraco Health Check: %0%</key>
-
-    <key alias="tlsHealthCheckSuccess">Votre site Web est configuré pour utiliser TLS version 1.2 ou supérieure pour les connexions sortantes.</key>
-    <key alias="tlsHealthCheckWarn">Les connexions sortantes de votre site Web sont servies via un ancien protocole. Vous devriez envisager de le mettre à jour pour utiliser TLS 1.2 ou supérieur.</key>
   </area>
   <area alias="redirectUrls">
     <key alias="disableUrlTracker">Désactiver URL tracker</key>
